+++ Properties/AssemblyInfo.cs
@@ -5,10 +5,10 @@ using System.Runtime.Versioning;
 
 [assembly: AssemblyCompany("cooooked")]
 [assembly: AssemblyConfiguration("Release")]
-[assembly: AssemblyFileVersion("4.1.143.0")]
-[assembly: AssemblyInformationalVersion("4.1.143")]
+[assembly: AssemblyFileVersion("4.1.146.0")]
+[assembly: AssemblyInformationalVersion("4.1.146")]
 [assembly: AssemblyProduct("ViewLab")]
 [assembly: AssemblyTitle("xr-viewlab")]
 [assembly: TargetPlatform("Windows7.0")]
 [assembly: SupportedOSPlatform("Windows7.0")]
-[assembly: AssemblyVersion("4.1.143.0")]
+[assembly: AssemblyVersion("4.1.146.0")]